% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohe_multi.R
\name{ohe_multi_multiple}
\alias{ohe_multi_multiple}
\title{ワンホットエンコーディング（複数列）}
\usage{
ohe_multi_multiple(data, columns)
}
\arguments{
\item{data}{データフレーム（ワンホットエンコーディングを行う対象のデータ）。}

\item{columns}{ワンホットエンコーディングを行う列名のベクトル（文字列）。}
}
\value{
元のデータフレームにワンホットエンコーディングされた列を追加したデータフレーム。
}
\description{
\code{ohe_multi_multiple} 指定した複数列に対してワンホットエンコーディングを一括で適用します。
}
\examples{
data <- tibble(ID = 1:3, col1 = c("A,B", "B,C", "A"), col2 = c("X,Y", "Y,Z", "X"))
ohe_multi_multiple(data, c("col1", "col2"))
}
