% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohe_multi.R
\name{ohe_multi_simple}
\alias{ohe_multi_simple}
\title{One-Hot Encoding (Single Column)}
\usage{
ohe_multi_simple(data, column_name)
}
\arguments{
\item{data}{A data frame containing the target data for one-hot encoding.}

\item{column_name}{A string specifying the column name to perform one-hot encoding on.}
}
\value{
A data frame with the original data and additional one-hot encoded columns.
}
\description{
\code{ohe_multi_simple} performs one-hot encoding (OHE) by splitting a specified column into unique values.
}
\examples{
data <- tibble(ID = 1:3, col = c("A,B", "B,C", "A"))
ohe_multi_simple(data, "col")
}
