% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohe_multi.R
\name{ohe_multi_multiple}
\alias{ohe_multi_multiple}
\title{One-Hot Encoding (Multiple Columns)}
\usage{
ohe_multi_multiple(data, columns)
}
\arguments{
\item{data}{A data frame containing the target data for one-hot encoding.}

\item{columns}{A vector of strings specifying the column names to perform one-hot encoding on.}
}
\value{
A data frame with the original data and additional one-hot encoded columns.
}
\description{
\code{ohe_multi_multiple} applies one-hot encoding (OHE) to multiple specified columns at once.
}
\examples{
data <- tibble(ID = 1:3, col1 = c("A,B", "B,C", "A"), col2 = c("X,Y", "Y,Z", "X"))
ohe_multi_multiple(data, c("col1", "col2"))
}
