% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohe_multi.R
\name{ohe_multi_simple}
\alias{ohe_multi_simple}
\title{ワンホットエンコーディング（単一列）}
\usage{
ohe_multi_simple(data, column_name)
}
\arguments{
\item{data}{データフレーム（ワンホットエンコーディングを行う対象のデータ）。}

\item{column_name}{ワンホットエンコーディングを行う列の名前（文字列）。}
}
\value{
元のデータフレームにワンホットエンコーディングされた列を追加したデータフレーム。
}
\description{
\code{ohe_multi_simple} 指定した列をユニークな値ごとに分割し、ワンホットエンコーディング（OHE）を実施
}
\examples{
data <- tibble(ID = 1:3, col = c("A,B", "B,C", "A"))
ohe_multi_simple(data, "col")
}
